// src/app/models/user-complemetn.ts
export interface UserComplemetnInterfase {
    id?: string;
    name?: string;
    lastName?: string;
    email?: string;
    documentType?: string;
    document?: number;
    phone?: number;
    address?: string;
    birthday?: string;
    weight?: number;
    height?: number;
    bloodType?: string;
    rh?: string;

}